Rename shared_components to design_system and add components and foundation subfolders
