+++ example/widgetbook.dart
@@ -1,8 +1,8 @@
-import 'package:concordium_wallet/shared_components/account_summary_card/account_summary_card.dart';
-import 'package:concordium_wallet/shared_components/account_summary_card/card_decorations.dart';
+import 'package:concordium_wallet/design_system/components/account_summary_card/account_summary_card.dart';
+import 'package:concordium_wallet/design_system/components/account_summary_card/card_decorations.dart';
 import 'package:flutter/material.dart';
 import 'package:widgetbook/widgetbook.dart';
-import 'package:concordium_wallet/shared_components/button.dart';
+import 'package:concordium_wallet/design_system/components/buttons/button.dart';
 
 import 'components/button_material.dart';
 import 'components/button_generated.dart';
